refatoração...
A piece of code should be where
you expect to find it — and, if not, you should re-factor to get it there. Clean code
